# Optimize opcode check