Hello,

First I'd like to thank you for this great project. I am teaching the basics of embedded systems and this could be a great resource for me in the future !

While I was taking a look at the source code I came along this quick optimization in the instruction code. I hope I didn't miss something obvious.

Simply replacing `if` with `else if` provides a welcomed performance boost.
All these `if` on every opcode adds up, so adding `else if` reduces the amount of checks per opcode.

However I liked the clear opcode separation by a line but the prettier doesn't allow this with `else if` so it is a bit ugly and condensed now.

I am not too familiar with hardware simulators but my guess would be that use of a map of opcodes    or switch statements would provide even greater benefits.

With this PR the last instruction still requires evaluation of all the preceding `if` and `else if`.

## Benchmarks

Before this PR:
```
Benchmark: cpu-benchmark
  avrInstruction: 66.145 ns(-117%)
  avrInstructionObjMap: 30.535 ns(0%)
  avrInstructionUintArray: 50.937 ns(-67%)
```

With this PR:
```
Benchmark: cpu-benchmark
  avrInstruction: 47.811 ns(-58%)
  avrInstructionObjMap: 30.208 ns(0%)
  avrInstructionUintArray: 51.428 ns(-70%)
```