@urish 2020-03-18
Thanks for taking the time to create this pull-request Glenn! This seems to be a simple yet clever way to improve the performance of the simulator, a sort of quick-win.

Have you tried comparing the speed of actual code (such as the blink example from the demo project)? What speed did you get and what browser/device have you used?