@gfeun 2020-03-18
Yes, i was just editing my PR to add that, you're too fast :smile:

So i observed a ~7% difference using the demo after 5s of simulation time

Without PR: 
Simulation time: 00:05.093 (27%)

With PR:
Simulation time: 00:05.218 (34%)