@gfeun 2020-03-18
Oh, I wasn't expecting such massive difference between Firefox and Chromium performance.

Just tested on Chromium 80

Without PR:
Simulation time: 00:10.218 (83%)

With PR:
Simulation time: 00:10.375 (137%)