@urish 2020-03-18
Good job Glenn! I just released this as part of 0.6.1. 

Where do you teach embedded systems? 

> I am not too familiar with hardware simulators but my guess would be that use of a map of opcodes or switch statements would provide even greater benefits.

Yeah, that's the purpose of the `benchmark` folder - I have started it in effort to figure out how to further improve performance. If you look at `benchmark/index.ts`, it checks exactly that - comparing the performance of `avrInstruction` v.s. using a map / UInt16Array to decode the instructions in one go. 

If you are interested in further looking in these directions let me know, we can brainstorm together.