@gfeun 2020-03-18
Thank you for your reactivity !

I teach at a French university called Université de Bretagne Occidentale. It is not my main job, I only intervene as a substitute teacher for practical work on an introductory course. Practical work include teaching assembly basics, C and the use of some µController peripherals (GPIO, ADC, 7-segment display, LED, timers, buzzer). We teach using AVR libraries without Arduino abstractions to give an understanding of the low level interactions with the hardware.

What is of interest to me is that we have custom boards with all our peripherals and an atmega 2560 sitting in the middle. Since school has gone full remote we have to find ways to teach these concepts to students at home. I stumble upon your project and saw the great potential to be able to create a virtual board, or at least part of it (Button and LED would be a good starting point).

So I'm not an expert but I'll be glad to help !

I see that you have some interesting issues awaiting that may be easier for me to begin with than performance optimization :smile:
I'll take a few days to familiarize with Typescript and the codebase and will keep you updated.