@urish 2020-03-18
Thanks for the explanation, that's great to hear! 

Any help with the open issues will be blessed. Which peripherals do you have on your atmega2560 board?