@gfeun 2020-03-18
Our board look like this (this is our old one but peripherals are the same)
![image](https://i.imgur.com/GOZQb1S.jpg)

So we make use of GPIO for leds, buttons, 7-seg and switches
ADC for potentiometer and temperature sensor (Upper left)
Timers to generate signals for the buzzer
Usart for communication with a PC
Interupts for keypad

ADC would be a great next step and interesting to implement.