@urish 2020-03-18
👍  for the ADC!

As far as the other hardware, here is some stuff we already have: LEDs are there, 7-seg, push buttons, USART, and I also see that you have an LCD there, which we have too. 

So apart from ADC, the missing pieces are the keypad, buzzer , and switches.

If you want to check out the 7-seg, for instance, I created a small programming challenge that uses it, and would love to get your feedback on it:

[https://wokwi.com/challenge/seven-forty-two](https://wokwi.com/challenge/seven-forty-two?utm_source=github&utm_medium=social&utm_campaign=avr8js)