@gfeun 2020-03-18
Keypad is not necessary, switches shoudn't be hard and I saw that you have an early buzzer implementation in your simon game example. 
I'll dabble with some components and try to create an initial board webpage.

That's a good challenge and a great website.
I couldn't wrap my head around the 7-seg pinout, I'll try again tomorrow.