@urish 2020-03-19
Thanks for the feedback!

What do you think about building this board page on top of the wokwi.com platform? 

I have just released a beta editor that actually allows you to tinker with both the code and the hardware / explanations in the page, you can get it to by adding `?beta-editor=yeah` to any playground/challenge URL, e.g.:

[https://wokwi.com/playground/blink?beta-editor=yeah](https://wokwi.com/playground/blink?beta-editor=yeah&utm_source=github&utm_medium=social&utm_campaign=avr8js)

This editor allows you to preview your change immediately, but it currently does not support saving the code. Once if have some code you are happy with, you can send a pull-request to the [playgrounds](https://github.com/wokwi/wokwi-playgrounds) repo, and it will go live shortly after being merged.