@gfeun 2020-03-19
I like the possibility to write lessons using markdown while including electronic components using html ! That's super neat. The dynamic reload is great too.

~~I don't understand however how you link the components to specific pins ?~~ (found, with the **label** attribute, feels like magic :smile: )

I'd like to know, do you own wokwi ? Is it intended to be a commercial platform in the future ?

For the board I imagined an horizontally split window to fit the real board orientation. But this could be changed, i'll try to build a prototype.

A little feedback, I don't personally feel that the "comment section" is useful for this type of content. Imagining lots of comments here would distract from the lesson i think. 

It's a massive ecosystem of tools you built here, that's amazing.