@urish 2020-03-19
Great!

Yes, Wokwi is a project I started together with @AriellaE. The idea is to provide a platform where people can create Arduino, Electronics and IoT courses, both paid and free, and share them with the community.

Good call about the comments section. We actually added it pretty early in the project, and so far it has only been used by a single person to report about an issue. We'll probably remove it or make it optional (so that they playground owner can decide if they want to have comments or not).

Thanks :-)